% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_tool_registry.R
\name{clear_tools}
\alias{clear_tools}
\title{Clear all registered tools}
\usage{
clear_tools()
}
\description{
Clear all registered tools
}
\seealso{
\code{\link[=list_tools]{list_tools()}}, \code{\link[=register_tool]{register_tool()}}, \code{\link[=unregister_tool]{unregister_tool()}}, \code{\link[=get_tool]{get_tool()}}
}
