% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_tool_registry.R
\name{get_tool}
\alias{get_tool}
\title{Get a tool by name}
\usage{
get_tool(name)
}
\arguments{
\item{name}{tool name}
}
\description{
Get a tool by name
}
\seealso{
\code{\link[=register_tool]{register_tool()}}, \code{\link[=unregister_tool]{unregister_tool()}}, \code{\link[=list_tools]{list_tools()}}
}
