% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_memory.R
\name{new_buffer_memory}
\alias{new_buffer_memory}
\title{Create a new BufferMemory}
\usage{
new_buffer_memory(n = 10)
}
\arguments{
\item{n}{Number of most recent messages to keep.}
}
\value{
Buffer memory object.
}
\description{
Create a new BufferMemory
}
\examples{
mem <- new_buffer_memory(3)
mem$add(create_message("user","hi"))
}
