% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_tool_registry.R
\name{register_tool}
\alias{register_tool}
\title{Register a tool globally}
\usage{
register_tool(tool)
}
\arguments{
\item{tool}{A \code{llmr_tool}.}
}
\description{
Register a tool globally
}
\seealso{
\code{\link[=unregister_tool]{unregister_tool()}}, \code{\link[=list_tools]{list_tools()}}, \code{\link[=get_tool]{get_tool()}}, \link{tool-api}
}
